Fix type warning from ide.
Former-commit-id: f3ff783552551c4dd04bcd03d6073ae40a69a0a9

+++ analysis/EventAnalysis.cc
@@ -74,7 +74,7 @@ EventAnalysis::EventAnalysis(Event*  eventIn,
 
 void EventAnalysis::SetPrintModuloFraction(double fraction)
 {
-  int    nEntries = chain->GetEntries();
+  int nEntries = (int)chain->GetEntries();
   printModulo = (int)ceil(nEntries * fraction);
   if (printModulo <= 0)
     {printModulo = 1;}
